pyproject.toml
[tool.poetry]
name = "fastapi-demo"
version = "0.1.0"
description = ""
authors = ["Wenbao Xu <mail@xuwenbao.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.5"
sqlmodel = "^0.0.22"
loguru = "^0.7.2"
pyyaml = "^6.0.2"
uvicorn = "^0.32.0"
pydantic-settings = "^2.6.1"
pymysql = "^1.1.1"
aiomysql = "^0.2.0"
cryptography = "^43.0.3"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
